chore: improve linting rules, fix Dart version mismatch, add global error handling, and enhance CI pipeline
- Add 14 custom lint rules (prefer_const_constructors, avoid_print, directives_ordering, etc.)
- Apply 211 auto-fixes across 65 files
- Fix Dart SDK constraint (^3.10.4 -> >=3.10.0 <4.0.0) to match local Dart 3.11.4
- Update CI Flutter version from 3.24.0 to 3.41.0
- Add flutter analyze and flutter test steps to CI pipeline
- Add global error handling (FlutterError.onError, PlatformDispatcher.onError, ErrorWidget.builder)
- Sort pub dependencies alphabetically
- Fix widget test to handle ErrorWidget.builder override

+++ lib/screens/practice/practice_hub_screen.dart
@@ -1,18 +1,18 @@
 import 'package:flutter/material.dart';
-import "../../models/practice_activity.dart";
-import '../../models/practice_activity.dart';
+
 import '../../logic/practice_service.dart';
 import '../../logic/star_service.dart';
-import "../../widgets/adaptive_practice_card.dart";
-import '../../widgets/responsive_container.dart';
-import '../../utils/responsive.dart';
+import '../../models/practice_activity.dart';
 import '../../theme/app_colors.dart';
-import 'spelling_practice_screen.dart';
+import '../../utils/responsive.dart';
+import '../../widgets/adaptive_practice_card.dart';
+import '../../widgets/responsive_container.dart';
 import 'listening_practice_screen.dart';
-import 'speed_match_screen.dart';
 import 'memory_game_screen.dart';
-import 'pronunciation_practice_screen.dart';
 import 'phrase_practice_screen.dart';
+import 'pronunciation_practice_screen.dart';
+import 'speed_match_screen.dart';
+import 'spelling_practice_screen.dart';
 
 /// Pantalla principal del hub de prácticas
 class PracticeHubScreen extends StatefulWidget {
@@ -307,7 +307,7 @@ class _PracticeHubScreenState extends State<PracticeHubScreen> {
     final totalCount = _activities.where((a) => a.isUnlocked).length;
     final practiceStars = _progressMap.values.fold(
       0,
-      (sum, p) => sum + (p.starsEarned as int),
+      (sum, p) => sum + (p.starsEarned),
     );
 
     return Container(
